Replace asm with intrinsics
The `asm!` directive didn't compile on nightly anymore as it was. The
new code still has issues, most obviously that Bit Test and Set
instruction is not available.

Another issue is the fact that most of these are available through
stable atomics interfaces, so those should be used instead. However,
they are not available for u128, even though they were for a short time:

https://github.com/rust-lang/rust/issues/39590

Lastly, memory-ordering is not specified. However, this is not much
different from what it used to be like.